Rewrote all queries to either Criteria with sql restriction or filters on collections

+++ whynot-core/src/main/java/nl/ru/cmbi/whynot/hibernate/GenericDAO.java
@@ -68,7 +68,7 @@ public interface GenericDAO<T, ID extends Serializable> {
 
 		List<Entry> getUnannotated(Databank db);
 
-		long getPresentCount(Databank db);
+		int getPresentCount(Databank db);
 
 		long getValidCount(Databank db);
